fix: address generation lifecycle review
