+++ crates/openai-frontend/src/chat.rs
@@ -6,13 +6,11 @@ use serde_json::Value;
 use crate::{
     common::{
         AgentSessionIdentity, FinishReason, PromptCacheRetention, ReasoningConfig, ReasoningEffort,
-        StopSequence, StreamOptions, Usage, completion_id, now_unix_secs,
+        StopSequence, StreamOptions, Usage, agent_session_metadata, agent_session_source_metadata,
+        completion_id, now_unix_secs, set_agent_session_metadata,
     },
     errors::OpenAiError,
 };
-
-const INTERNAL_AGENT_SESSION_ID: &str = "mesh_internal_agent_session_id";
-const INTERNAL_AGENT_SESSION_SOURCE: &str = "mesh_internal_agent_session_source";
 
 #[derive(Debug, Clone, Deserialize, PartialEq)]
 pub struct ChatCompletionRequest {
@@ -49,32 +47,17 @@ pub struct ChatCompletionRequest {
 
 impl ChatCompletionRequest {
     pub(crate) fn set_agent_session(&mut self, identity: Option<AgentSessionIdentity>) {
-        self.extra.remove(INTERNAL_AGENT_SESSION_ID);
-        self.extra.remove(INTERNAL_AGENT_SESSION_SOURCE);
-        if let Some(identity) = identity {
-            self.extra.insert(
-                INTERNAL_AGENT_SESSION_SOURCE.to_owned(),
-                Value::String(identity.source().label().to_owned()),
-            );
-            self.extra.insert(
-                INTERNAL_AGENT_SESSION_ID.to_owned(),
-                Value::String(identity.id().to_owned()),
-            );
-        }
+        set_agent_session_metadata(&mut self.extra, identity);
     }
 
     #[must_use]
     pub fn agent_session(&self) -> Option<&str> {
-        self.extra
-            .get(INTERNAL_AGENT_SESSION_ID)
-            .and_then(Value::as_str)
+        agent_session_metadata(&self.extra)
     }
 
     #[must_use]
     pub fn agent_session_source(&self) -> Option<&str> {
-        self.extra
-            .get(INTERNAL_AGENT_SESSION_SOURCE)
-            .and_then(Value::as_str)
+        agent_session_source_metadata(&self.extra)
     }
 
     pub fn effective_max_tokens(&self) -> Option<u32> {
