+++ crates/skippy-server/src/frontend/local_generation/token_generation.rs
@@ -1,7 +1,5 @@
-use crate::frontend::LinearProposalDisposition;
 use crate::frontend::NativeMtpDecodeOptions;
 use crate::frontend::NativeMtpDraft;
-use crate::frontend::NativeMtpDraftOrigin;
 use crate::frontend::NativeMtpVerifier;
 use crate::frontend::generation::GenerationCacheStats;
 use crate::frontend::generation::LocalGeneration;
@@ -9,15 +7,8 @@ use crate::frontend::generation::OpenAiGenerationIds;
 use crate::frontend::generation::PhaseTimer;
 use crate::frontend::generation::StageOpenAiBackend;
 use crate::frontend::generation::TokenControl;
-use crate::frontend::generation::decode_token_phase;
-use crate::frontend::generation_receipt::{
-    GenerationReceiptObservation, GenerationStart, complete_generation_before_cleanup,
-};
-use crate::frontend::linear_proposal::{
-    LinearProposalDiscardReason, LinearProposalExecutionParams, LinearProposalQueryOutcome,
-    LinearProposalQueryParams, execute_linear_proposal_with_terminal_discard,
-    greedy_linear_proposal_admitted, query_linear_proposal, report_linear_proposal_receipt,
-};
+use crate::frontend::generation_receipt::{GenerationStart, complete_generation_before_cleanup};
+use crate::frontend::linear_proposal::greedy_linear_proposal_admitted;
 use crate::frontend::util::openai_backend_error;
 use crate::frontend::util::saturating_u32;
 use crate::kv_integration::KvStageIntegration;
@@ -33,6 +24,7 @@ use skippy_runtime::NativeMtpDraft as RuntimeNativeMtpDraft;
 use skippy_runtime::SamplingConfig;
 use std::cell::RefCell;
 use std::collections::BTreeMap;
+use std::sync::Arc;
 use std::time::Duration;
 
 use super::{LocalGenerationReceiptFinalization, prompt_fits_single_prefill_sample};
@@ -52,30 +44,30 @@ struct KvRecordResult {
     proactive_eviction_error: Option<anyhow::Error>,
 }
 
-struct DecodeState {
-    decoded_tokens: usize,
-    current: i32,
-    stopped: bool,
-    runtime_lock_wait_ms: f64,
-    runtime_lock_wait_max_ms: f64,
-    runtime_lock_hold_ms: f64,
-    runtime_lock_hold_max_ms: f64,
-    runtime_lock_acquires: usize,
-    runtime_sessions_before: Option<RuntimeSessionStats>,
-    runtime_sessions_after: Option<RuntimeSessionStats>,
-    hook_request: Option<ChatCompletionRequest>,
-    hook_runtime: Option<tokio::runtime::Handle>,
-    generation_hooks_active: bool,
-    linear_proposal_max_tokens: usize,
-    linear_context_tokens: Option<Vec<i32>>,
-    emit_token_debug: bool,
-    native_mtp_options: NativeMtpDecodeOptions,
-    native_mtp: NativeMtpVerifier,
-    post_prefill_hook_checked: bool,
-    last_mid_generation_hook_at: Option<usize>,
+pub(super) struct DecodeState {
+    pub(super) decoded_tokens: usize,
+    pub(super) current: i32,
+    pub(super) stopped: bool,
+    pub(super) runtime_lock_wait_ms: f64,
+    pub(super) runtime_lock_wait_max_ms: f64,
+    pub(super) runtime_lock_hold_ms: f64,
+    pub(super) runtime_lock_hold_max_ms: f64,
+    pub(super) runtime_lock_acquires: usize,
+    pub(super) runtime_sessions_before: Option<RuntimeSessionStats>,
+    pub(super) runtime_sessions_after: Option<RuntimeSessionStats>,
+    pub(super) hook_request: Option<ChatCompletionRequest>,
+    pub(super) hook_runtime: Option<tokio::runtime::Handle>,
+    pub(super) generation_hooks_active: bool,
+    pub(super) linear_proposal_max_tokens: usize,
+    pub(super) linear_context_tokens: Option<Vec<i32>>,
+    pub(super) emit_token_debug: bool,
+    pub(super) native_mtp_options: NativeMtpDecodeOptions,
+    pub(super) native_mtp: NativeMtpVerifier,
+    pub(super) post_prefill_hook_checked: bool,
+    pub(super) last_mid_generation_hook_at: Option<usize>,
 }
 
-enum LinearProposalProgress {
+pub(super) enum LinearProposalProgress {
     NotUsed,
     Continue,
     Stop,
@@ -90,20 +82,29 @@ impl StageOpenAiBackend {
         let session_id = request.ids.session_label.clone();
         let receipt_request_id = request.ids.request_id;
         let receipt_session_id = request.ids.session_id;
-        let receipt_prompt_token_ids = request.prompt_token_ids;
+        let receipt_prompt_token_ids = self
+            .generation_receipt
+            .as_ref()
+            .map(|_| Arc::<[i32]>::from(request.prompt_token_ids));
         if let Some(config) = self.generation_receipt.as_ref() {
             config.begin(GenerationStart {
                 request_id: receipt_request_id,
                 session_id: receipt_session_id,
                 agent_session_id: request.ids.agent_session_id.clone(),
-                prompt_token_ids: receipt_prompt_token_ids.to_vec().into_boxed_slice(),
+                prompt_token_ids: Arc::clone(
+                    receipt_prompt_token_ids
+                        .as_ref()
+                        .expect("receipt prompt exists when receipt config exists"),
+                ),
             });
         }
-        let receipt_observation = self.generation_receipt.as_ref().map(|_| {
-            RefCell::new(Some(GenerationReceiptObservation::new(
-                usize::try_from(request.max_tokens)
-                    .expect("supported targets represent u32 token budgets as usize"),
-            )))
+        let receipt_observation = self.generation_receipt.as_ref().map(|config| {
+            RefCell::new(Some(
+                config.observation(
+                    usize::try_from(request.max_tokens)
+                        .expect("supported targets represent u32 token budgets as usize"),
+                ),
+            ))
         });
         let mut receipt_cancelled = false;
         let mut receipt_model_generation_elapsed = None;
@@ -112,7 +113,7 @@ impl StageOpenAiBackend {
             if let Some(observation) = receipt_observation.as_ref()
                 && let Some(observation) = observation.borrow_mut().as_mut()
             {
-                observation.record_token(token_id, request.ids.request_started_at.elapsed())?;
+                observation.record_token(token_id, request.ids.request_started_at.elapsed());
             }
             let control = on_token(token_id)?;
             if control == TokenControl::Stop
@@ -154,7 +155,7 @@ impl StageOpenAiBackend {
                         request_id: receipt_request_id,
                         session_id: receipt_session_id,
                         agent_session_id: request.ids.agent_session_id.as_deref(),
-                        prompt_token_ids: receipt_prompt_token_ids,
+                        prompt_token_ids: receipt_prompt_token_ids.unwrap_or_default(),
                         observation: receipt_observation,
                         cancelled: receipt_cancelled,
                         model_generation_elapsed: receipt_model_generation_elapsed,
@@ -829,395 +830,9 @@ impl StageOpenAiBackend {
         }
         self.emit_decode_summary(request, &mut state, cache_stats, decode_timer)
     }
-
-    fn try_execute_linear_proposal(
-        &self,
-        request: &LocalGeneration<'_>,
-        session_id: &str,
-        state: &mut DecodeState,
-        emit_token: &mut impl FnMut(i32) -> OpenAiResult<TokenControl>,
-    ) -> OpenAiResult<LinearProposalProgress> {
-        let Some(config) = self.linear_proposal_ingress.as_ref() else {
-            return Ok(LinearProposalProgress::NotUsed);
-        };
-        let Some(committed_token_ids) = state.linear_context_tokens.as_mut() else {
-            return Ok(LinearProposalProgress::NotUsed);
-        };
-        let remaining_new_tokens =
-            (request.max_tokens as usize).saturating_sub(state.decoded_tokens);
-        // Prefill leaves the final prompt token undecoded. When whole-prompt
-        // prefill also samples the first target token, decoded_tokens is one;
-        // otherwise it is zero. Those two modes therefore share this position.
-        let base_position = u64::try_from(
-            request
-                .prompt_token_ids
-                .len()
-                .saturating_sub(1)
-                .checked_add(state.decoded_tokens)
-                .ok_or_else(|| {
-                    OpenAiError::backend("linear proposal base position exceeds usize")
-                })?,
-        )
-        .map_err(|_| OpenAiError::backend("linear proposal base position exceeds u64"))?;
-        let queried = match query_linear_proposal(
-            config,
-            LinearProposalQueryParams {
-                request_id: request.ids.request_id,
-                session_id: request.ids.session_id,
-                prompt_token_count: request.prompt_token_ids.len(),
-                decode_step: state.decoded_tokens,
-                committed_token_count: request
-                    .prompt_token_ids
-                    .len()
-                    .saturating_add(state.decoded_tokens),
-                remaining_new_tokens,
-                runtime_max_proposal_tokens: state.linear_proposal_max_tokens,
-            },
-        )? {
-            LinearProposalQueryOutcome::NoProposal => None,
-            LinearProposalQueryOutcome::DeadlineExceeded {
-                proposal_elapsed_us,
-            } => {
-                let mut attrs = BTreeMap::new();
-                attrs.insert(
-                    "llama_stage.linear_proposal.discard_reason".to_string(),
-                    json!("deadline_exceeded"),
-                );
-                attrs.insert(
-                    "llama_stage.linear_proposal.proposal_us".to_string(),
-                    json!(proposal_elapsed_us),
-                );
-                self.telemetry
-                    .emit("stage.openai_linear_proposal_late", attrs);
-                None
-            }
-            LinearProposalQueryOutcome::Ready(queried) => Some(queried),
-        };
-        let Some(queried) = queried else {
-            return Ok(LinearProposalProgress::NotUsed);
-        };
-        let decision_id = queried.proposal.decision_id.clone();
-        let receipt = execute_linear_proposal_with_terminal_discard(config, &decision_id, || {
-            self.execute_local_linear_proposal(
-                LinearProposalExecutionParams {
-                    session_id,
-                    current: state.current,
-                    base_position,
-                    generated_len: state.decoded_tokens,
-                    max_new_tokens: request.max_tokens as usize,
-                    sampling: request.sampling,
-                    chat_sampling_metadata: request.chat_sampling_metadata,
-                    prompt_token_count: request.prompt_token_ids.len(),
-                },
-                queried,
-                emit_token,
-            )
-        })?;
-        if receipt.is_none() {
-            let discard_failed = config
-                .source()
-                .discard(&decision_id, LinearProposalDiscardReason::PositionMismatch)
-                .is_err();
-            if discard_failed {
-                self.telemetry.emit(
-                    "stage.openai_linear_proposal_discard_failed",
-                    BTreeMap::from([(
-                        "llama_stage.linear_proposal.discard_reason".to_string(),
-                        json!("position_mismatch"),
-                    )]),
-                );
-            }
-        }
-        let Some(receipt) = receipt else {
-            return Ok(LinearProposalProgress::NotUsed);
-        };
-        if report_linear_proposal_receipt(config, &receipt).is_some() {
-            let mut attrs = BTreeMap::new();
-            receipt.insert_telemetry_attrs(&mut attrs);
-            attrs.insert(
-                "llama_stage.linear_proposal.report_outcome".to_string(),
-                json!("failed"),
-            );
-            self.telemetry
-                .emit("stage.openai_linear_proposal_report_failed", attrs);
-        }
-        let proposal_runtime_lock_wait_ms =
-            Duration::from_micros(receipt.runtime_lock_wait_us).as_secs_f64() * 1_000.0;
-        let proposal_runtime_lock_hold_ms =
-            Duration::from_micros(receipt.runtime_lock_hold_us).as_secs_f64() * 1_000.0;
-        state.runtime_lock_wait_ms += proposal_runtime_lock_wait_ms;
-        state.runtime_lock_wait_max_ms = state
-            .runtime_lock_wait_max_ms
-            .max(proposal_runtime_lock_wait_ms);
-        state.runtime_lock_hold_ms += proposal_runtime_lock_hold_ms;
-        state.runtime_lock_hold_max_ms = state
-            .runtime_lock_hold_max_ms
-            .max(proposal_runtime_lock_hold_ms);
-        state.runtime_lock_acquires = state
-            .runtime_lock_acquires
-            .saturating_add(receipt.runtime_lock_acquires);
-        state.decoded_tokens = state
-            .decoded_tokens
-            .checked_add(receipt.committed_tokens.len())
-            .ok_or_else(|| OpenAiError::backend("linear proposal decode count overflow"))?;
-        state.current = *receipt
-            .committed_tokens
-            .last()
-            .ok_or_else(|| OpenAiError::backend("linear proposal receipt committed no tokens"))?;
-        committed_token_ids.extend_from_slice(&receipt.committed_tokens);
-        let stopped_by_proposal = receipt.disposition == LinearProposalDisposition::Stopped;
-        if state.emit_token_debug {
-            let mut proposal_attrs = BTreeMap::new();
-            receipt.insert_telemetry_attrs(&mut proposal_attrs);
-            self.telemetry
-                .emit_debug("stage.openai_linear_proposal", proposal_attrs);
-        }
-        if stopped_by_proposal || state.decoded_tokens >= request.max_tokens as usize {
-            Ok(LinearProposalProgress::Stop)
-        } else {
-            Ok(LinearProposalProgress::Continue)
-        }
-    }
-
-    fn decode_one_token(
-        &self,
-        request: &LocalGeneration<'_>,
-        session_id: &str,
-        state: &mut DecodeState,
-        emit_token: &mut impl FnMut(i32) -> OpenAiResult<TokenControl>,
-    ) -> OpenAiResult<TokenControl> {
-        let decode_step = state.decoded_tokens;
-        let token_timer = PhaseTimer::start();
-        let decode_call_timer = PhaseTimer::start();
-        let (
-            predicted,
-            mut native_mtp_draft,
-            token_batch_size,
-            token_batch_wait_ms,
-            token_runtime_lock_wait_ms,
-            token_runtime_lock_hold_ms,
-        ) = if request.native_mtp_enabled {
-            let lock_timer = PhaseTimer::start();
-            let mut runtime = self
-                .runtime
-                .lock()
-                .map_err(|_| OpenAiError::backend("runtime lock poisoned"))?;
-            let token_runtime_lock_wait_ms = lock_timer.elapsed_ms();
-            let hold_timer = PhaseTimer::start();
-            let (predicted, draft) = decode_native_mtp(
-                &mut *runtime,
-                session_id,
-                state.current,
-                request.sampling.enabled.then_some(request.sampling),
-                state.native_mtp_options.max_draft_tokens,
-            )?;
-            let native_mtp_draft = draft;
-            let token_batch_size = 1;
-            let token_batch_wait_ms = 0.0;
-            let token_runtime_lock_hold_ms = hold_timer.elapsed_ms();
-            (
-                predicted,
-                native_mtp_draft,
-                token_batch_size,
-                token_batch_wait_ms,
-                token_runtime_lock_wait_ms,
-                token_runtime_lock_hold_ms,
-            )
-        } else {
-            let outcome = self.decode_batcher.decode(
-                session_id,
-                state.current,
-                request.sampling.enabled.then_some(request.sampling),
-            )?;
-            (
-                outcome.predicted,
-                None,
-                outcome.batch_size,
-                outcome.batch_wait_ms,
-                outcome.runtime_lock_wait_ms,
-                outcome.runtime_lock_hold_ms,
-            )
-        };
-        state.current = predicted;
-        if native_mtp_draft
-            .as_ref()
-            .is_some_and(|draft: &NativeMtpDraft| {
-                draft.tokens.len() < state.native_mtp_options.min_draft_tokens
-            })
-        {
-            native_mtp_draft = None;
-        }
-        let is_first_draft = state.decoded_tokens == 0;
-        let draft_origin = if is_first_draft {
-            NativeMtpDraftOrigin::InitialSerial
-        } else {
-            NativeMtpDraftOrigin::SerialAfterGap
-        };
-        let native_mtp_decision = request.native_mtp_enabled.then(|| {
-            state
-                .native_mtp
-                .observe_target_token(state.current, 0, native_mtp_draft, draft_origin)
-        });
-        state.runtime_lock_wait_ms += token_runtime_lock_wait_ms;
-        state.runtime_lock_wait_max_ms = state
-            .runtime_lock_wait_max_ms
-            .max(token_runtime_lock_wait_ms);
-        state.runtime_lock_hold_ms += token_runtime_lock_hold_ms;
-        state.runtime_lock_hold_max_ms = state
-            .runtime_lock_hold_max_ms
-            .max(token_runtime_lock_hold_ms);
-        state.runtime_lock_acquires += 1;
-        let token_decode_ms = if state.emit_token_debug {
-            decode_call_timer.elapsed_ms()
-        } else {
-            0.0
-        };
-        let (token_signal, signal_window, token_signal_ms) = if state.generation_hooks_active {
-            let signal_timer = PhaseTimer::start();
-            let mut runtime = self
-                .runtime
-                .lock()
-                .map_err(|_| OpenAiError::backend("runtime lock poisoned"))?;
-            state
-                .runtime_sessions_before
-                .get_or_insert_with(|| runtime.session_stats());
-            let token_signal = runtime.last_token_signal(session_id).ok();
-            let signal_window = runtime.signal_window(session_id, 16).ok();
-            state.runtime_sessions_after = Some(runtime.session_stats());
-            (token_signal, signal_window, signal_timer.elapsed_ms())
-        } else {
-            (None, None, 0.0)
-        };
-        let injected_current = if state.generation_hooks_active {
-            self.maybe_run_generation_hooks(
-                session_id,
-                &mut state.hook_request,
-                state.hook_runtime.as_ref(),
-                state.decoded_tokens,
-                &mut state.post_prefill_hook_checked,
-                &mut state.last_mid_generation_hook_at,
-                token_signal,
-                signal_window,
-            )?
-        } else {
-            None
-        };
-        if let Some(injected_current) = injected_current {
-            state.current = injected_current;
-            return Ok(TokenControl::Continue);
-        }
-        state.decoded_tokens += 1;
-        if let Some(committed_token_ids) = state.linear_context_tokens.as_mut() {
-            committed_token_ids.push(state.current);
-        }
-        if state.emit_token_debug {
-            let mut token_attrs = self.openai_attrs(request.ids);
-            token_attrs.insert("llama_stage.decode_step".to_string(), json!(decode_step));
-            token_attrs.insert(
-                "llama_stage.decode_token_phase".to_string(),
-                json!(decode_token_phase(
-                    u32::try_from(decode_step).unwrap_or(u32::MAX)
-                )),
-            );
-            token_attrs.insert(
-                "llama_stage.stage0_compute_ms".to_string(),
-                json!(token_timer.elapsed_ms()),
-            );
-            token_attrs.insert(
-                "llama_stage.decode_call_ms".to_string(),
-                json!(token_decode_ms),
-            );
-            token_attrs.insert(
-                "llama_stage.decode_batch_size".to_string(),
-                json!(token_batch_size),
-            );
-            token_attrs.insert(
-                "llama_stage.decode_batch_wait_ms".to_string(),
-                json!(token_batch_wait_ms),
-            );
-            token_attrs.insert("llama_stage.signal_ms".to_string(), json!(token_signal_ms));
-            token_attrs.insert(
-                "llama_stage.runtime_lock_wait_ms".to_string(),
-                json!(token_runtime_lock_wait_ms),
-            );
-            token_attrs.insert(
-                "llama_stage.runtime_lock_hold_ms".to_string(),
-                json!(token_runtime_lock_hold_ms),
-            );
-            token_attrs.insert(
-                "llama_stage.predicted_token".to_string(),
-                json!(state.current),
-            );
-            if let Some(native_mtp_decision) = native_mtp_decision {
-                token_attrs.insert(
-                    "llama_stage.native_mtp.verification".to_string(),
-                    json!(native_mtp_decision.label()),
-                );
-            }
-            token_attrs.insert("llama_stage.message_kind".to_string(), json!("DecodeToken"));
-            self.emit_openai_phase("stage.openai_decode_token", token_timer, token_attrs);
-        }
-        emit_token(state.current)
-    }
-
-    fn emit_decode_summary(
-        &self,
-        request: &LocalGeneration<'_>,
-        state: &mut DecodeState,
-        cache_stats: &mut GenerationCacheStats,
-        decode_timer: PhaseTimer,
-    ) -> OpenAiResult<Duration> {
-        let mut attrs = self.openai_attrs(request.ids);
-        attrs.insert(
-            "llama_stage.decode_token_count".to_string(),
-            json!(state.decoded_tokens),
-        );
-        attrs.insert(
-            "llama_stage.runtime_lock_wait_ms".to_string(),
-            json!(state.runtime_lock_wait_ms),
-        );
-        attrs.insert(
-            "llama_stage.runtime_lock_wait_max_ms".to_string(),
-            json!(state.runtime_lock_wait_max_ms),
-        );
-        attrs.insert(
-            "llama_stage.runtime_lock_hold_ms".to_string(),
-            json!(state.runtime_lock_hold_ms),
-        );
-        attrs.insert(
-            "llama_stage.runtime_lock_hold_max_ms".to_string(),
-            json!(state.runtime_lock_hold_max_ms),
-        );
-        attrs.insert(
-            "llama_stage.runtime_lock_acquires".to_string(),
-            json!(state.runtime_lock_acquires),
-        );
-        if let Some(stats) = state.runtime_sessions_before.as_ref() {
-            Self::insert_runtime_session_stats(
-                &mut attrs,
-                "llama_stage.runtime_sessions_before",
-                stats,
-            );
-        }
-        if let Some(stats) = state.runtime_sessions_after.as_ref() {
-            Self::insert_runtime_session_stats(
-                &mut attrs,
-                "llama_stage.runtime_sessions_after",
-                stats,
-            );
-        }
-        request.speculative.insert_telemetry_attrs(&mut attrs);
-        let native_mtp_stats = state.native_mtp.stats();
-        cache_stats.native_mtp_stats = native_mtp_stats;
-        let model_generation_elapsed = decode_timer.start_instant.elapsed();
-        cache_stats.predicted_ms = model_generation_elapsed.as_secs_f64() * 1_000.0;
-        native_mtp_stats.insert_attrs(&mut attrs);
-        self.emit_openai_summary("stage.openai_decode", decode_timer, attrs);
-        Ok(model_generation_elapsed)
-    }
 }
 
-trait NativeMtpRuntime {
+pub(super) trait NativeMtpRuntime {
     fn decode_sampled_mtp(
         &mut self,
         session_id: &str,
@@ -1270,7 +885,7 @@ impl NativeMtpRuntime for RuntimeState {
     }
 }
 
-fn decode_native_mtp(
+pub(super) fn decode_native_mtp(
     runtime: &mut impl NativeMtpRuntime,
     session_id: &str,
     token_id: i32,
