+++ crates/skippy-server/src/lib.rs
@@ -30,11 +30,10 @@ pub use frontend::{
     CONTEXT_BUDGET_MAX_TOKENS, DEFAULT_EMBEDDED_MAX_TOKENS, EmbeddedOpenAiArgs,
     EmbeddedOpenAiBackend, EmbeddedOpenAiRequestDefaults, EmbeddedReasoningBudget,
     EmbeddedReasoningEnabled, EmbeddedReasoningFormat, LinearProposal, LinearProposalDiscardReason,
-    LinearProposalDisposition, LinearProposalIngress, LinearProposalIngressConfig,
-    LinearProposalQuery, LinearProposalReceipt, NativeMtpProposalConfig, NgramExtensionConfig,
-    NgramProposalConfig, NgramProposerKind, OpaqueProposalDecisionId, OpenAiGuardrailsConfig,
-    OpenAiGuardrailsStatus, OpenAiGuardrailsTarget, SpeculativeDecodeConfig, VerifyWindowConfig,
-    embedded_openai_backend,
+    LinearProposalDisposition, LinearProposalIngress, LinearProposalQuery, LinearProposalReceipt,
+    NativeMtpProposalConfig, NgramExtensionConfig, NgramProposalConfig, NgramProposerKind,
+    OpaqueProposalDecisionId, OpenAiGuardrailsConfig, OpenAiGuardrailsStatus,
+    OpenAiGuardrailsTarget, SpeculativeDecodeConfig, VerifyWindowConfig, embedded_openai_backend,
 };
 pub use skippy_protocol::StageConfig;
 pub use tokenizer::{MAX_TOKENIZE_TOKENS, TokenizerCapability, TokenizerCapabilityError};
